Иконки созданы отдельными компонентами
